django_examples/store/requirements.txt
alembic==0.9.8
banal==0.3.3
certifi==2018.1.18
chardet==3.0.4
click==6.7
dataset==1.0.6
decorator==4.2.1
Django==2.0.2
djangorestframework==3.7.7
Flask==0.12.2
flask-marshmallow==0.8.0
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
httpie==0.9.9
idna==2.6
ipdb==0.11
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.0
marshmallow-sqlalchemy==0.13.2
model-mommy==1.5.1
normality==0.5.11
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2018.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
